+++ pkg/config/tlscfg/flags.go
@@ -22,14 +22,12 @@ import (
 
 const (
 	tlsPrefix         = ".tls"
-	tlsEnabledOld     = tlsPrefix
 	tlsEnabled        = tlsPrefix + ".enabled"
 	tlsCA             = tlsPrefix + ".ca"
 	tlsCert           = tlsPrefix + ".cert"
 	tlsKey            = tlsPrefix + ".key"
 	tlsServerName     = tlsPrefix + ".server-name"
 	tlsClientCA       = tlsPrefix + ".client-ca"
-	tlsClientCAOld    = tlsPrefix + ".client.ca"
 	tlsSkipHostVerify = tlsPrefix + ".skip-host-verify"
 )
 
@@ -51,7 +49,6 @@ type ServerFlagsConfig struct {
 func (c ClientFlagsConfig) AddFlags(flags *flag.FlagSet) {
 	if c.ShowEnabled {
 		flags.Bool(c.Prefix+tlsEnabled, false, "Enable TLS when talking to the remote server(s)")
-		flags.Bool(c.Prefix+tlsEnabledOld, false, "(deprecated) see --"+c.Prefix+tlsEnabled)
 	}
 	flags.String(c.Prefix+tlsCA, "", "Path to a TLS CA (Certification Authority) file used to verify the remote server(s) (by default will use the system truststore)")
 	flags.String(c.Prefix+tlsCert, "", "Path to a TLS Certificate file, used to identify this process to the remote server(s)")
@@ -66,12 +63,10 @@ func (c ClientFlagsConfig) AddFlags(flags *flag.FlagSet) {
 func (c ServerFlagsConfig) AddFlags(flags *flag.FlagSet) {
 	if c.ShowEnabled {
 		flags.Bool(c.Prefix+tlsEnabled, false, "Enable TLS on the server")
-		flags.Bool(c.Prefix+tlsEnabledOld, false, "(deprecated) see --"+c.Prefix+tlsEnabled)
 	}
 	flags.String(c.Prefix+tlsCert, "", "Path to a TLS Certificate file, used to identify this server to clients")
 	flags.String(c.Prefix+tlsKey, "", "Path to a TLS Private Key file, used to identify this server to clients")
 	flags.String(c.Prefix+tlsClientCA, "", "Path to a TLS CA (Certification Authority) file used to verify certificates presented by clients (if unset, all clients are permitted)")
-	flags.String(c.Prefix+tlsClientCAOld, "", "(deprecated) see --"+c.Prefix+tlsClientCA)
 }
 
 // InitFromViper creates tls.Config populated with values retrieved from Viper.
@@ -79,10 +74,6 @@ func (c ClientFlagsConfig) InitFromViper(v *viper.Viper) Options {
 	var p Options
 	if c.ShowEnabled {
 		p.Enabled = v.GetBool(c.Prefix + tlsEnabled)
-
-		if !p.Enabled {
-			p.Enabled = v.GetBool(c.Prefix + tlsEnabledOld)
-		}
 	}
 	p.CAPath = v.GetString(c.Prefix + tlsCA)
 	p.CertPath = v.GetString(c.Prefix + tlsCert)
@@ -99,19 +90,11 @@ func (c ServerFlagsConfig) InitFromViper(v *viper.Viper) Options {
 	var p Options
 	if c.ShowEnabled {
 		p.Enabled = v.GetBool(c.Prefix + tlsEnabled)
-
-		if !p.Enabled {
-			p.Enabled = v.GetBool(c.Prefix + tlsEnabledOld)
-		}
 	}
 	p.CertPath = v.GetString(c.Prefix + tlsCert)
 	p.KeyPath = v.GetString(c.Prefix + tlsKey)
 	if c.ShowClientCA {
 		p.ClientCAPath = v.GetString(c.Prefix + tlsClientCA)
-		if s := v.GetString(c.Prefix + tlsClientCAOld); s != "" {
-			// using legacy flag
-			p.ClientCAPath = s
-		}
 	}
 	return p
 }